added error handling

+++ milkshake/src/main/java/com/wildcodeschool/milkshake/controller/RecetteController.java
@@ -3,9 +3,8 @@ package com.wildcodeschool.milkshake.controller;
 import com.wildcodeschool.milkshake.entity.Recette;
 import com.wildcodeschool.milkshake.service.RecetteService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 import java.util.Optional;
 
 @RestController
@@ -19,27 +18,60 @@ public class RecetteController {
     }
 
     @GetMapping()
-    public List<Recette> findAll() {
-        return this.recetteService.findAll();
+    public ResponseEntity<?> findAll() {
+        try {
+            return ResponseEntity.ok(this.recetteService.findAll());
+        } catch(RuntimeException e) {
+            return ResponseEntity.status(500).body(e.getMessage());
+        }
     }
 
     @GetMapping("/{id}")
-    public Optional<Recette> findById(@PathVariable long id){
-        return this.recetteService.findById(id);
+    public ResponseEntity<?> findById(@PathVariable long id){
+        try{
+            Optional<Recette> recipe = this.recetteService.findById(id);
+            if (recipe.isEmpty()) return ResponseEntity.status(404).body("Not found");
+
+            return ResponseEntity.ok(recipe.get());
+        } catch(RuntimeException e){
+            return ResponseEntity.status(500).body(e.getMessage());
+        }
     }
 
     @PostMapping()
-    public void save(@RequestBody Recette recette){
-        this.recetteService.saveOrUpdate(recette, null);
+    public ResponseEntity<?> save(@RequestBody Recette recette){
+        try{
+            this.recetteService.saveOrUpdate(recette, null);
+            return ResponseEntity.ok("Recette saved successfully");
+        } catch(RuntimeException e){
+            return ResponseEntity.status(500).body(e.getMessage());
+        }
     }
 
    @PutMapping("/{id}")
-   public void update(@RequestBody Recette recette, @PathVariable long id){
-        this.recetteService.saveOrUpdate(recette, id);
+   public ResponseEntity<?> update(@RequestBody Recette recette, @PathVariable long id){
+        try{
+            if (this.recetteService.findById(id).isEmpty())
+                return ResponseEntity.status(404).body("Not found");
+
+            this.recetteService.saveOrUpdate(recette, id);
+            return ResponseEntity.ok("Recette updated successfully");
+        } catch(RuntimeException e) {
+            return ResponseEntity.status(500).body(e.getMessage());
+        }
+
    }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable long id){
-        this.recetteService.deleteById(id);
+    public ResponseEntity<?> deleteById(@PathVariable long id){
+        try{
+            if (this.recetteService.findById(id).isEmpty())
+                return ResponseEntity.status(404).body("Not found");
+
+            this.recetteService.deleteById(id);
+            return ResponseEntity.ok().body("Recipe deleted successfully");
+        } catch(RuntimeException e) {
+            return ResponseEntity.status(500).body(e.getMessage());
+        }
     }
 }
